refactor: collapse submodule
Very few functions left after refactor.
